### 2022-11-08 #4918 @editorialbot
[PRE REVIEW]: Water Systems Integrated Modelling framework, WSIMOD: A Python package for integrated modelling of water quality and quantity across the water cycle
**Submitting author:** @barneydobson (<a href="http://orcid.org/0000-0002-0149-4124">Barnaby Dobson</a>)
**Repository:** https://github.com/barneydobson/wsi
**Branch with paper.md** (empty if default branch):
**Version:** 0.2_joss_submission
