### 2022-11-08 #4918 @barneydobson
Thanks!

Potential reviewers selected from the provided list, based on 'hydrology'
keyword search and a quick skim of relevant papers:
jlarsen-usgs: ***@***.***
alessandroamaranto: ***@***.***
cheginit: ***@***.***
dvalters: (no email)

Additional relevant reviewers who have a strong background in research
software:
Jon Herman - Jon Herman | Center for Watershed Sciences (ucdavis.edu)
<https://watershed.ucdavis.edu/people/jon-herman>
Leah Jackson-Blake - Leah Amber Jackson-Blake - NIVA
<https://www.niva.no/en/Employees/leah-amber-jackson-blake>
Gregor Burger - gregorburger (Gregor Burger) (github.com)
<https://github.com/gregorburger>

Best,
Barney

On Tue, 8 Nov 2022 at 15:52, The Open Journals editorial robot <
***@***.***> wrote:

> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
>
>
> OK DOIs
>
>
>
> - None
>
>
>
> MISSING DOIs
>
>
>
> - 10.31223/osf.io/64cvn may be a valid DOI for title: Protecting rivers by integrating supply-wastewater infrastructure planning and coordinating operational decisions
>
> - 10.1016/j.scitotenv.2021.150642 may be a valid DOI for title: Hierarchical systems integration for coordinated urban-rural water quality management at a catchment scale
>
> - 10.3389/frwa.2021.641462 may be a valid DOI for title: Integrated modelling to support analysis of COVID-19 impacts on London’s water system and in-river water quality
>
> - 10.1002/essoar.10507590.1 may be a valid DOI for title: A reduced complexity model with graph partitioning for rapid hydraulic assessment of sewer networks
>
> - 10.3389/frwa.2021.773974 may be a valid DOI for title: The value of aggregated city scale models to rapidly assess SuDS in combined sewer systems
>
> - 10.1002/essoar.10511984.1 may be a valid DOI for title: Optimisation of Urban-Rural Nature-Based Solutions for Integrated Catchment Water Management
>
>
>
> INVALID DOIs
>
>
>
> - None
>
>
