### 2022-11-09 #4918 @barneydobson
Thanks for the info Kristen.

There were a couple of preprints that the bot picked up instead of the
proper dois:
10.1088/1748-9326/abb050 should be for "Protecting rivers by integrating
supply-wastewater infrastructure planning and coordinating operational
decisions"
10.1029/2021WR030778 should be for "A reduced complexity model with graph
partitioning for rapid hydraulic assessment of sewer networks"

I've also added the doi entry in the bibliography file on the repository
too in case there are any further issues.

Best,
Barney
