### 2022-11-21 812d90d
Update land.py
Better handling of pollution on land
