### 2022-12-06 #4918 @barneydobson
Thanks Chris,

Potential reviewers selected from the provided list, based on 'hydrology' keyword search and a quick skim of relevant papers:
jlarsen-usgs: ***@***.******@***.***>
alessandroamaranto: ***@***.******@***.***>
cheginit: ***@***.******@***.***>
dvalters: (no email)

Additional relevant reviewers who have a strong background in research software:
Jon Herman - Jon Herman | Center for Watershed Sciences (ucdavis.edu)<https://watershed.ucdavis.edu/people/jon-herman>
Leah Jackson-Blake - Leah Amber Jackson-Blake - NIVA<https://www.niva.no/en/Employees/leah-amber-jackson-blake>
Gregor Burger - gregorburger (Gregor Burger) (github.com)<https://github.com/gregorburger>

Best
Barney

Get Outlook for Android<https://aka.ms/AAb9ysg>
