### 2022-12-07 #4918 @crvernon
:wave: Hi there @cheginit!  Are you up for reviewing this submission?  As a reminder, we do checklist based reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
