### 2022-12-08 #4996 @editorialbot
[REVIEW]: Water Systems Integrated Modelling framework, WSIMOD: A Python package for integrated modelling of water quality and quantity across the water cycle
**Submitting author:** @barneydobson (<a href="http://orcid.org/0000-0002-0149-4124">Barnaby Dobson</a>)
**Repository:** https://github.com/barneydobson/wsi
**Branch with paper.md** (empty if default branch):
**Version:** 0.3_joss_reviewed
