### 2022-12-08 #4996 @crvernon
👋 @barneydobson @cheginit @jlarsen-usgs This is the review thread for the paper. All of our communications will happen here from now on.
