### 2023-01-06 4014275
Update wtw.py
Fix WW reuse behaviour
