+++ docs/demo/scripts/cherwell_demo_catchwat_residencetank.py
@@ -1,0 +1,309 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Dec 30 09:56:02 2022
+
+@author: leyan
+"""
+
+
+from wsimod.nodes.waste import Waste
+from wsimod.nodes.land import Land
+from wsimod.nodes.nodes import Node
+from wsimod.nodes.wtw import WWTW
+from wsimod.nodes.storage import River
+from wsimod.nodes.discharge_point import Discharge_point
+from wsimod.nodes.storage import Groundwater
+from wsimod.orchestration.model import Model
+from wsimod.arcs.arcs import Arc
+from wsimod.core import constants
+from wsimod.demo.create_oxford import create_timeseries
+import os
+import pandas as pd
+from matplotlib import pyplot as plt
+from datetime import datetime
+
+data_folder= os.path.join("D:\\", "Huawei", "PhD", "Year4", "cherwell_demo")
+
+precipitation = pd.read_csv(os.path.join(data_folder, "Cherwell, Thame and Wye_rainfall_1975_2018_day_25_km.csv"))
+precipitation.columns = ['date', 'value']
+precipitation['variable'] = 'precipitation'
+precipitation['value'] *= constants.MM_TO_M
+et0 = pd.read_csv(os.path.join(data_folder, "reference_ET.csv"))
+et0.columns = ['date', 'value']
+et0['variable'] = 'et0'
+et0['value'] *= constants.MM_TO_M
+tasmax = pd.read_csv(os.path.join(data_folder, "Cherwell, Thame and Wye_tasmax_1975_2018_day_25_km.csv")).set_index('date')
+tasmin = pd.read_csv(os.path.join(data_folder, "Cherwell, Thame and Wye_tasmin_1975_2018_day_25_km.csv")).set_index('date')
+temperature = (tasmax['tasmax'] + tasmin['tasmin']) / 2
+temperature.name = 'value'
+temperature = pd.DataFrame(temperature)
+temperature['variable'] = 'temperature'
+temperature = temperature.reset_index(drop=False)
+
+input_data = pd.concat([precipitation, et0, temperature])
+input_data.date = pd.to_datetime(input_data.date)
+land_inputs = input_data.set_index(['variable','date']).value.to_dict()
+dates = input_data.date.drop_duplicates()
+
+constants.set_default_pollutants()
+
+crop_para = pd.read_csv(os.path.join(data_folder, "Crop_parameters.csv")).set_index('lucode')
+fer = pd.read_csv(os.path.join(data_folder, "fer_timeseries.csv")).set_index('date')
+man = pd.read_csv(os.path.join(data_folder, "man_timeseries.csv")).set_index('date')
+surfaces = []
+for i in range(0, len(crop_para.index)):
+    
+    idx = crop_para.index[i]
+    
+    area = crop_para.loc[idx, 'area'] * constants.KM2_TO_M2
+    kc_ini = crop_para.loc[idx, 'Kc_ini']
+    kc_mid = crop_para.loc[idx, 'Kc_mid']
+    kc_end = crop_para.loc[idx, 'Kc_end']
+    cc_s1 = datetime.strptime('2001-'+crop_para.loc[idx, 'cc_s1'], '%Y-%m-%d').timetuple().tm_yday
+    cc_s2 = datetime.strptime('2001-'+crop_para.loc[idx, 'cc_s2'], '%Y-%m-%d').timetuple().tm_yday
+    cc_s3 = datetime.strptime('2001-'+crop_para.loc[idx, 'cc_s3'], '%Y-%m-%d').timetuple().tm_yday
+    cc_s4 = datetime.strptime('2001-'+crop_para.loc[idx, 'cc_s4'], '%Y-%m-%d').timetuple().tm_yday
[…]
