+++ docs/demo/scripts/oxford_demo.py
@@ -87,8 +87,7 @@ from matplotlib import pyplot as plt
 # %% [markdown]
 # Load input data
 # %%
-data_folder= os.path.join(os.path.abspath(''),
-                               "docs","demo","data")
+data_folder= os.path.join(os.path.abspath(os.path.join(os.path.dirname("__file__"),os.path.pardir)), "data")
 
 #Use this path if opening in jupyter
 # data_folder = os.path.join(os.path.split(os.path.abspath(''))[0],"data") 
