### 2023-01-23 barneydobson/wsi#6 @jlarsen-usgs
JOSS Review
@barneydobson

I am beginning my review on WSIMOD for your JOSS publication and I will be tracking my review comments here and editing this issue as they come up, I should be finished doing my review by the end of this week. Feel free to address them all within this issue, or convert each of the comments to their own issues.

- [x] General reviews
   - [x] I see that you have some tests set up for WSIMOD. It would be great to see these tests incorporated into a continuous integration system, such as github actions.
   - [x] In your WSI docs you have a detailed summary and statement of need. Can you adapt a concise version of this text into your Readme.md file. The Welcome WSIMOD section would be a good place to add this information.
   - [x] Add a requirements section in your Readme.md. What versions of python does your code support? What python packages are required or optionally required to use WSIMOD.
   - [x]  Community guidelines: A code of conduct, contributing guidelines, and either an issue template or a short statement about where to seek support for the software should be added to the reposity.
   - [x] Add installation instructions to readme.md.
   - [x] Update installation instructions in your api documentation (and in readme.md) to show installation with pip (`pip install .`). It'd be nice to add instructions on how to install directly from github for those users who aren't likely to dig into the source code (`pip install https://github.com/barneydobson/wsi/archive/refs/heads/main.zip`)

- [x] paper reviews:
   - [x] The statement of need should expand on the gaps in scientific code/research this software is trying to fill and clarify what this software does and does not do.
       - [x] How does this hydrologic simulator compare to other existing modeling software? What gaps does it fill?
   - [x] Lines 22 and 23 state that you simulate "all parts of the water cycle". This should be replaced with a more detailed description of the processes that you simulate.
   - [x]  Please expand on the capabilities and limitations of your hydrologic simulator. I think it'd add a lot to the paper if you explained your approach of Arcs and Nodes to simulate fluxes and sinks/storage.
   - [x] Optional: A short example application could really help illustrate the functionality of WSIMOD. I'd really consider consisely presenting one of your example problems (1-2 paragraphs and a figure) to illustrate the utility of the WSIMOD.

- [x] code reviews:
   - [x] the demo problems for your documentation are currently missing data and not reproducible. All of the example problems are missing "timeseries_data.csv". Until this is fixed I can't really finish my review.
   - [x] the demo problems are currently set up to generate your documentation tutorials. These could be adapted into either example scripts or interactive jupyter notebooks (in addition to keeping the current docs workflow) where the user does not have to reset the `data_folder` path variable to access the correct path to your data.
   - [x] the bulk density of soil in your `GrowingSurface` module seems low to me (1300 kg/m3), The average bulk denity of soil is around 2650 kg/m3.
