### 2023-01-24 barneydobson/wsi#7 @cheginit
JOSS Review
I am one of the reviewers of your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4996). I am going to include my comments and suggestions in this issue and will edit until all issues are resolved.

Comments on the accompanying paper:
- [x] The paper is missing some important sections:
  * [x] No discussion on the current state of the field and the knowledge gap that this package is aiming to fill.
  * [x] The discussion on the software capabilities are very limited. You need to provide more details on high-level functionalities of the software, such as, Urban processes that it can model, required input data, number of calibrated vs non-calibrated parameters, and its applicability range and limitations.
- [x] Line 42 should be removed.

Comment on the package:
- [x Suggestions (non-blockers):
  * [x] Including a (long) TOC in the README is not helpful. It'd be better if it includes contents similar to the [About](https://barneydobson.github.io/wsi/paper/paper/) section of the software's website and some figures demonstrating its capabilities. Essentially, it should give users a quick glance to the content of the repository.
  * [x] Since the software includes a test suite, it's easy to set up a testing GitHub Actions workflow.
  * [x] Using linters could help improve the code quality and catch possible errors. For example, [pre-commit](https://pre-commit.com/) can be a good starting point.
  * [x] In my opinion, the way that input data are structured and users have to make all components manually is not user-friendly at all. Since there are many components that required user input, using a config file is much more user-friendly. Many modeling packages follow this input structure, such as SWMM. It's easy to implement such a config file-based input using `yaml` files in Python.
- [x] Issues (blockers):
  * [x] There is no clear and direct guidance or discussion in the documentation regarding the required input data and pre-processing them in a form that the model accepts. There are some hints in the examples, but there should be a dedicated section for processing raw input data from different formats to the required format.
  * [x] Considering that this package combines several models, at least a brief discussion on their assumptions and governing equations are required so a user can make informed decisions regarding the necessary components that need to be included for getting acceptable results. A very detailed and technical discussion is not required like a journal paper, but at least a user should be able to have a good understanding of the model assumptions by reading the documentation.
