### 2023-01-25 7f31913
Update README
-Include small summary in both README.md and on docs index page
-Include more detailed install instructions in both README.md and docs installation page
-Reduce table of contents in README.md
-Include optional requirements in README.md
