+++ docs/index.md
@@ -3,6 +3,22 @@
 This is the documentation for the [WSIMOD project](https://github.com/barneydobson/wsi).
 WSIMOD stands for the Water Systems Integrated Modelling framework.
 
+The terrestrial water cycle is a highly interconnected system where the 
+movement of water is affected by physical and human processes. Thus, 
+environmental models may become inaccurate if they do not provide a complete 
+picture of the water cycle, missing out on unexpected opportunities and 
+omitting impacts that arise from complex interactions. WSIMOD is a modelling 
+framework to integrate these different processes. It provides a message passing 
+interface to enable different subsystem models to communicate water flux and 
+water quality information between each other, and self-contained 
+representations of the key parts of the water cycle (rivers, reservoirs, urban 
+and rural hydrological catchments, treatment plants, and pipe networks).
+We created WSIMOD to enable a user greater flexibility in setting up their 
+water cycle models, motivated by the abundance of non-textbook water systems 
+that we have experienced in industry collaboration. The WSIMOD Python package 
+provides tutorials and examples to help modellers create nodes, connect them 
+with arcs, and create simulations.
+
 ## Table Of Contents
 
 The documentation follows the best practice for
