### 2023-01-25 64160eb
Update requirements
-Remove unused packages
-Move majority of packages into optional
-Replace to_datetime to avoid need for pandas dependency to carry out unit testing
