+++ tests/test_wtw.py
@@ -16,8 +16,6 @@ from wsimod.nodes.wtw import WTW, WWTW, FWTW
 from wsimod.nodes.nodes import Node
 from wsimod.nodes.sewer import Sewer
 from wsimod.arcs.arcs import Arc
-from pandas import to_datetime
-
     
 
 
