+++ .github/workflows/ci.yml
@@ -1,0 +1,18 @@
+name: CI
+
+on: [push]
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python
+      uses: actions/setup-python@v2
+      with:
+        python-version: ["3.7", "3.8", "3.9", "3.10"]
+    - name: Install and build
+      run: pip install .
+    - name: Run tests
+      run: pytest
