### 2023-01-25 672af89
Update ci.yml
fix CI to Python3.10 only
