+++ .github/workflows/ci.yml
@@ -11,7 +11,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v2
       with:
-        python-version: ["3.7", "3.8", "3.9", "3.10"]
+        python-version: 3.10
     - name: Install and build
       run: pip install .
     - name: Run tests
