### 2023-01-26 barneydobson/wsi#6 @barneydobson
Thanks @jlarsen-usgs these are incredibly helpful comments - I'm working on all of these but to just reply about the `data_folder` issue, it seems to me to be on the repo in this folder (https://github.com/barneydobson/wsi/tree/main/docs/demo/data/processed).

Possibly I need to make it more clear, but there is a comment that says when inputting the `data_folder`, that a different address should be used when running in interactive Python (e.g., line 51 in https://github.com/barneydobson/wsi/blob/main/docs/demo/scripts/quickstart_demo.py)

As part of my revisions, I will make this the default `data_folder` address - but I am letting you know now so that you can continue looking at them.
