### 2023-01-26 2d4ee52
Update demo data paths
Default to useable data_folder, rather than the documentation address
