+++ docs/demo/scripts/oxford_demo.py
@@ -87,11 +87,12 @@ from matplotlib import pyplot as plt
 # %% [markdown]
 # Load input data
 # %%
+
+#Use this path if compiling documentation
 data_folder= os.path.join(os.path.abspath(''),
                                "docs","demo","data")
-
-#Use this path if opening in jupyter
-# data_folder = os.path.join(os.path.split(os.path.abspath(''))[0],"data") 
+#Use this path otherwise
+data_folder = os.path.join(os.path.split(os.path.abspath(''))[0],"data") 
 
 input_fid = os.path.join(data_folder, "processed", "timeseries_data.csv")
 input_data = pd.read_csv(input_fid)
