+++ docs/demo/scripts/quickstart_demo.py
@@ -46,10 +46,11 @@ from matplotlib import pyplot as plt
 # evapotranspiration (et0).
 # %%
 
+#Use this path if compiling documentation
 data_folder= os.path.join(os.path.abspath(''),
                                "docs","demo","data")
-# data_folder = os.path.join(os.path.split(os.path.abspath(''))[0],"data") #Use this path if opening in jupyter
-
+#Use this path otherwise
+data_folder = os.path.join(os.path.split(os.path.abspath(''))[0],"data") 
 
 input_fid = os.path.join(data_folder, "processed", "timeseries_data.csv")
 input_data = pd.read_csv(input_fid)
