### 2023-01-26 barneydobson/wsi#6 @jlarsen-usgs
Thanks @barneydobson, I see that the file is in the WSIMOD repository. I'll update my copy of the repo to make sure I have the most recent version and retry later today.
