### 2023-01-26 bd62b84
Introduce model saving/loading from a config file
-YAML based config file load/save functions now part of the model object (including csv read/write to avoid pandas dependency, function to identify what parameters are needed to initialise a node/arc, and the load/save functions themselves, which carry out the expected processes and have some exceptions for surfaces and data writing)
-Required custom datetime object to avoid dependency on pandas (small edits to tests, new datetime object in model)
-Required that all initialisation parameters are preserved in a node/arc object (explicit definiti
[…]
