+++ wsimod/orchestration/model.py
@@ -14,8 +14,63 @@ from wsimod.nodes.nodes import QueueTank, Tank, Node
 import os
 os.environ['USE_PYGEOS'] = '0'
 import sys
-
+import inspect
+import csv
+import gzip
+import yaml
 from math import log10
+from datetime import datetime
+
+class to_datetime():
+    def __init__(self, date_string):
+        self._date = self._parse_date(date_string)
+        
+    def __str__(self):
+        return self._date.strftime("%Y-%m-%d")
+    
+    @property
+    def dayofyear(self):
+        return self._date.timetuple().tm_yday
+    
+    @property
+    def day(self):
+        return self._date.day
+    
+    @property
+    def year(self):
+        return self._date.year
+    
+    @property
+    def month(self):
+        return self._date.month
+    
+    def to_period(self, args = 'M'):
+        return f"{self._date.year}-{str(self._date.month).zfill(2)}"    
+    
+    def is_leap_year(self):
+        year = self._date.year
+        return year % 4 == 0 and (year % 100 != 0 or year % 400 == 0)
+    
+    def _parse_date(self, date_string, date_format = "%Y-%m-%d %H:%M:%S"):
+        try:
+            return datetime.strptime(date_string, date_format)
+        except ValueError:
+            try:
+                return datetime.strptime(date_string, "%Y-%m-%d")
+            except ValueError:
+                try:
+                    return datetime.strptime(date_string, "%Y-%m")
+                except ValueError:
+                    raise ValueError
+                    
+    def __eq__(self, other):
+        if isinstance(other, to_datetime):
+            return self._date == other._date
+        return False
+
+    def __hash__(self):
+        return hash(self._date)
+    
 class Model(WSIObj):
     def __init__(self):
         """Object to contain nodes and arcs that provides a default
@@ -38,6 +93,144 @@ class Model(WSIObj):
         self.nodes_type = set(getattr(nodes,x)(name='').__class__.__name__ for x in self.nodes_type).union(['Foul'])
         self.nodes_type = {x : {} for x in self.nodes_type}
         
+    def get_init_args(self,cls):
+        """
+        Get the arguments of the __init__ method for a class and its superclasses
+        """
+        init_args = []
+        for c in cls.__mro__:
+            # Get the arguments of the __init__ method
+            args = inspect.getfullargspec(c.__init__).args[1:]
+            init_args.extend(args)
+        return init_args
+
+    def open_func(self, file_path, mode):
+        if mode == "rt" and file_path.endswith(".gz"):
+            return gzip.open(file_path, mode)
+        else:
+            return open(file_path, mode)
+        
+    def read_csv(self, file_path, delimiter=","):
+        with self.open_func(file_path, "rt") as f:
+            reader = csv.DictReader(f, delimiter=delimiter)
+            data = {}
+            for row in reader:
+                key = (row['variable'], to_datetime(row['time']))
+                value = float(row['value'])
+                data[key] = value
+            return data
+
+    def write_csv(self, data, fixed_data, filename, compress = False):
+        if compress:
+            open_func = gzip.open
+            mode = 'wt'
+        else:
+            open_func = open
+            mode = 'w'
+        with open_func(filename, mode, newline='') as csvfile:
+            writer = csv.writer(csvfile)
+            writer.writerow(list(fixed_data.keys()) + ['variable', 'time', 'value'])
+            fixed_data_values = list(fixed_data.values())
+            for key, value in data.items():
+                writer.writerow(fixed_data_values + list(key) + [str(value)])
+    
+    def load(self, address, config_name = 'config.yml'):
+        
+        with open(os.path.join(address, config_name), "r") as file:
+            data = yaml.safe_load(file)
+        
+        nodes = data['nodes']
[…]
