### 2023-01-27 barneydobson/wsi#6 @barneydobson
Hi @jlarsen-usgs

Just providing a quick update on these tasks, note that I have not yet been able to meet with co-authors so I haven't focussed on paper edits. Changes are currently on the https://github.com/barneydobson/wsi/tree/joss_revisions branch and summarised below.
-Continuous integration included in a GitHub action.
-Concise summary added to Readme and will be added to Welcome page of documentation.
-Code of conduct added. Still figuring out the best approach to contributing guidelines/issue template.
-Requirements (these are only needed to run the demos) have now been added to the readme.
-Added install instructions to readme and edited according to your suggestion (but I might still iterate these to accommodate any changed requirements).
-Default data_folder updated RE previous conversation.
-I wanted to check the soil density with the person who's code the GrowingSurface module is based on.
