### 2023-01-27 barneydobson/wsi#7 @barneydobson
Hi, @cheginit. Thanks so much for these comments, they have been incredibly helpful. I just thought I would provide a short update on these tasks. Note that I have not yet been able to meet with co-authors so I haven't focussed on paper edits. Changes are currently on the https://github.com/barneydobson/wsi/tree/joss_revisions branch.
- TOC in readme edited according to suggestion.
- Testing added in a GitHub Actions workflow
- Linters I need to discuss with collaborators, I might create this as a GitHub issue to address in the future since it seems less urgent but would be quite a bit of work to address now.
- Agreed RE input data, this has been a super helpful suggestion. I have started doing exactly as you suggest (a `yaml` based config file that contains parameters and data addresses of input data). Though I still need to do further testing on this. I will aim to make a tutorial on this, but need to carefully consider what to include, because I think this could also address your comment on required input data/format.
- Agree with your comment RE guidance around input data, which I hope to address as above. I think generalising pre-processing will be too exhaustive a task, as there are many different modules included in WSIMOD, and each country/region/water company will have differently formatted raw data needed to inform each model element. Our intention was not for this WSIMOD package to be a geospatial data processing package (though we may aim to work on this in the future). If you have any detailed suggestions here though I would be grateful!
- I could possibly incorporate model assumptions/equations into the paper via a table format, though it will just be the assumptions/equations at this present time (I will discuss the merits of this with the co-authors). Perhaps preferable, and to ensure continuity without creating an unmanageable documentation maintanence burden, I will ensure that the docstrings for different model components list key assumptions and require in the contributing guidlines (which I will be adding on the suggestion of another reviewer) that new model components do the same.
