### 2023-01-27 38f5f33
Update install reqs for documentation
