### 2023-01-27 10c3cf9
Update model.py
More sensible save/load functionality (make data address relative to model address)
