### 2023-01-27 barneydobson/wsi#7 @cheginit
Thanks for starting to address the suggestions/issues.

Regarding the changes related to a user-friendly input, you can take your time since it's just a suggestion and not a publication blocker.

> I think generalising pre-processing will be too exhaustive a task, as there are many different modules included in WSIMOD, and each country/region/water company will have differently formatted raw data needed to inform each model element.

I agree with this, and that's not what I asked for. I was referring to adding clear discussion and guidance on the input data structure that WSIMOD accepts, such as variable names, units, and data types. Essentially, a user should be able to read this "input data processing" section of the documentation and transform their data to the format that WSIMOD can read.

> I will ensure that the docstrings for different model components list key assumptions and require in the contributing guidlines (which I will be adding on the suggestion of another reviewer) that new model components do the same.

This will be helpful, but docstring alone is not enough, a user should be able to read the software documentation and understand the key assumptions behind model components. This is necessary for community contribution. I didn't explicitly add a discussion on the missing contributing guidelines and code of conduct, since @jlarsen-usgs included it in #6.
