### 2023-01-29 8163da9
Update land.py
Add example assumption/parameter documentation
