+++ wsimod/nodes/land.py
@@ -50,9 +50,32 @@ class Land(Node):
                 dict and the second value is the time. Defaults to {}.
 
         Functions intended to call in orchestration:
-            run
-            
-            apply_irrigation (if used)
+            run  
+            apply_irrigation (if used)  
+        
+        Key assumptions:
+             - Percolation, surface runoff, and subsurface runoff, can be described with 
+                a residence-time method.
+             - Flows to percolation, surface runoff, and subsurface runoff are 
+                generated by different hydrological response units (subclasses 
+                of `land.py/Surface`), but aggregated for a given land node.
+             - Flows to percolation are distributed to `storage.py/Groundwater`
+                nodes while surface/subsurface runoff to `nodes.py/Node` or 
+                `storage.py/River` nodes.
+             - Input data associated with the land node (precipitation, 
+                temperature, evapotranspiartion) are the same for every surface.  
+             - Water received from `sewer.py/Sewer` objects is sent to the first
+                `land.py/ImperviousSurface` in the surfaces list.
+
+        Input data and parameter requirements:
+             - Precipitation and evapotranspiration are in the `data_input_dict`
+                at the model timestep.
+                _Units_: metres/timestep
+             - Temperature in the `data_input_dict` at the model timestep.
+                _Units_: C
+             - Residence time of surface, subsurface and percolation flows.
+                _Units_: number of timesteps
+
         """
         #Assign parameters
         self.subsurface_residence_time = subsurface_residence_time
@@ -87,10 +110,12 @@ class Land(Node):
 
             #Assign any irrigation functions 
             if isinstance(surfaces[-1], IrrigationSurface):
+                #TODO, this should probably be done in the Surface initialisation
                 self.irrigation_functions.append(surfaces[-1].irrigate)
             
             #Assign garden surface functions
             if isinstance(surfaces[-1], GardenSurface):
+                #TODO, this should probably be done in the Surface initialisation
                 self.push_check_handler[('Demand','Garden')] = surfaces[-1].calculate_irrigation_demand
                 self.push_set_handler[('Demand','Garden')] = surfaces[-1].receive_irrigation_demand
         
