+++ docs/tutorials.md
@@ -3,3 +3,4 @@
 1. [Quickstart](./demo/scripts/quickstart_demo.py)
 2. [WSIMOD model demonstration - Oxford](./demo/scripts/oxford_demo.py)
 3. [Land nodes - hydrology and agriculture](./demo/scripts/land_demo.py)
+4. [Model object - WSIMOD models](wsimod_models.md)
