+++ docs/wsimod_models.md
@@ -1,0 +1,218 @@
+# WSIMOD models
+
+1. [Introduction](#introduction)
+
+2. [What data do you need](#what-data-do-you-need)
+
+3. [WSIMOD model](#wsimod-model)
+
+    3.1. [Nodes](#nodes)
+
+    3.2. [Arcs](#arcs)
+
+    3.3. [Model properties](#model-properties)
+
+4. [Input data](#input-data)
+
+5. [Limitations](#limitations)
+
+## Introduction
+In other [tutorials](tutorials.md), we generally create nodes and arcs in a dictionary format
+to best explain how to understand what is going on in WSIMOD. However, this
+is not a streamlined way to setup or use a model in practice. Instead, we 
+provide a structured format that a model can be loaded from and the ability
+to save models to this format. In this tutorial we will describe the format, 
+and give examples.
+
+## What data do you need?
+WSIMOD combines a variety of different models into different types of 
+node/component. Our goal is a highly flexible approach to representing a wide 
+variety of water systems. The result is that your data requirements will be 
+highly specific to your project, the systems that you want to include, and the 
+questions you want to ask of your model. To understand what model components
+you should include, and what data requirements these have, we recommend 
+familiarising yourself with the WSIMOD approach through completing the 
+[tutorials](tutorials.md) and viewing the [API](reference.md). Ultimately though, 
+all nodes and arcs can be resolved in terms of dictionaries that
+determine their parameters/input data, which makes them ideal to be structured using 
+the [PyYAML](https://pyyaml.org/) data langauge. 
+
+## WSIMOD model
+As demonstrated in the tutorials, a [model object](./../reference-model/#wsimod.orchestration.model.Model) 
+is a helpful way to contain your nodes/arcs and orchestrate your simulation.
+From the model object, you can `save` or `load` data by providing a data 
+directory. Critically, these functions interact with a `config.yml` file that 
+contains all of the information to describe your model. Below, we will explain 
+this file using examples from [quickstart demo](./../demo/scripts/quickstart_demo), 
+highlighting the key features of this config file. You can create this yourself by 
+running the demo and appending the following command:
+
+```
+>>> my_model.save(<directory_address_on_your_filesystem>)
+```
+
+The top level entries of the `config.yml` file created in the provided directorty are: 
+```
+nodes
+arcs
+pollutants
+additive_pollutants
+non_additive_pollutants
+float_accuracy
+dates
+```
+, which cover all of the properties needed to describe a model object. 
+
+To load a model, we to initialise a model object and call the `load` function:
+
+```
+>>> my_model = Model()
+>>> my_model.load(<directory_address_on_your_filesystem>)
+```
+
+### Nodes
+The `nodes` entry of `config.yml` will contain the information required to initialise all of the components
+in your model. Below are a sewer node and default node. 
+
+```
+nodes:
+  my_sewer:
+    chamber_area: 1
+    name: my_sewer
+    pipe_time: 0
+    chamber_floor: 10
+    capacity: 0.04
+    pipe_timearea:
+      0: 1
+    type_: Sewer
+    node_type_override: Sewer
+  my_river:
+    name: my_river
+    type_: Node
+    node_type_override: Node
+```
+
+By inspecting the [`Sewer`](./../reference-sewer/#wsimod.nodes.sewer.Sewer) API, 
+we can see that this entry contains the parameters required to initiliase a 
+`Sewer` object, however the `capacity` and `name` fields have been updated to
+the values that they are set as in the [quickstart demo](./../demo/scripts/quickstart_demo).
[…]
