### 2023-01-30 642b1dc
Update core.py
typo in documentation
