### 2023-01-31 barneydobson/wsi#7 @barneydobson
@cheginit Just to double check with you before I implement it. I can add the key assumptions and input data to the model components' docstrings. See below an example for the Land node:

![image](https://user-images.githubusercontent.com/29261471/215750975-564a2214-5a4f-4647-b65b-fdc40b3265c5.png)

And I can write a script so that the documentation pulls all of these together on one page, so people have the ability to quickly identify what the model can do, what are its assumptions, and input data requirements. As I see it this would address both your blockers about input data guidance and key assumptions. But I wanted to check first before implementing it
