### 2023-01-31 barneydobson/wsi#7 @cheginit
Yes, this is very helpful and will address those issues.

You can also add citations. For example, you can check out one of my packages called [PyDaymet](https://docs.hyriver.io/autoapi/pydaymet/pet/index.html) and its associated lines in its [source code](https://github.com/hyriver/pydaymet/blob/8114660446f058a7ce74eeea68b495f6b8e32cec/pydaymet/pet.py#L679-L686), to see how referencing docstrings would look like. For the `method` parameter, I cite the paper that I used for the supported methods.
