+++ docs/paper/paper.bib
@@ -72,10 +72,48 @@
 }
 
 @article{Liu:2022b,
-  url = {https://www.essoar.org/doi/abs/10.1002/essoar.10511984.1},
-  doi={10.1002/essoar.10511984.1},
+  url = {https://www.sciencedirect.com/science/article/pii/S0301479722026184},
+  doi={10.1016/j.jenvman.2022.117045},
   title={Optimisation of Urban-Rural Nature-Based Solutions for Integrated Catchment Water Management},
   author={Liu, Leyang and Dobson, Barnaby and Mijic, Ana},
-  year={2022},
-  journal={ESSOAr: Earth and Space Science Open Archive},
+  year={2023},
+  journal={Journal of Environmental Management},
+}
+
+@article{Shrestha:2013,
+  url = {https://www.sciencedirect.com/science/article/pii/S136481521300114X},
+  doi = {10.1016/j.envsoft.2013.05.004},
+  title={OpenMI-based integrated sediment transport modelling of the river Zenne, Belgium},
+  author={Shrestha, Narayan Kumar and Leta, Olkeba Tolessa and De Fraine, Bruno and Van Griensven, Ann and Bauwens, Willy},
+  journal={Environmental Modelling \& Software},
+  volume={47},
+  pages={193--206},
+  year={2013},
+  publisher={Elsevier}
+}
+
+@article{Arnold:2012,
+  url = {https://elibrary.asabe.org/abstract.asp?aid=42256},
+  doi = {10.13031/2013.42259},
+  title={SWAT: Model use, calibration, and validation},
+  author={Arnold, Jeffrey G and Moriasi, Daniel N and Gassman, Philip W and Abbaspour, Karim C and White, Michael J and Srinivasan, Raghavan and Santhi, Chinnasamy and Harmel, RD and Van Griensven, Ann and Van Liew, Michael W and others},
+  journal={Transactions of the ASABE},
+  volume={55},
+  number={4},
+  pages={1491--1508},
+  year={2012},
+  publisher={American Society of Agricultural and Biological Engineers}
+}
+
+@article{Gironas:2010,
+  url = {https://www.sciencedirect.com/science/article/pii/S1364815209002989},
+  doi = {j.envsoft.2009.11.009},
+  title={A new applications manual for the Storm Water Management Model(SWMM)},
+  author={Giron{\'a}s, Jorge and Roesner, Larry A and Rossman, Lewis A and Davis, Jennifer},
+  journal={Environmental Modelling \& Software},
+  volume={25},
+  number={6},
+  pages={813--814},
+  year={2010},
+  publisher={Elsevier Science, P. O. Box 800 Kidlington Oxford OX 5 1 DX UK}
 }
