+++ docs/paper/paper.md
@@ -30,38 +30,71 @@ physical and human processes throughout. For example, rivers are a water
 supply, a receiver of wastewater, and an aggregate of many hydrological, 
 biological, and chemical processes. Thus, simulations of the water cycle that
 have highly constrained boundaries may miss key interactions that create 
-unanticipated impacts or unexpected opportunities [@Dobson:2020;@Liu:2022]. 
+unanticipated impacts or unexpected opportunities[@Dobson:2020;@Liu:2022]. 
 Integrated environmental models aim to resolve the issue of boundary 
-conditions, however they have some key limitations [@Rauch:2017], and in 
+conditions, however they have some key limitations[@Rauch:2017], and in 
 particular we find a significant need for a parsimonious, self-contained suite 
 that is accessible and easy to setup. 
 
 # Statement of need
 
-WSIMOD is a Python package for integrated modelling of the water cycle. It 
+Traditional approaches to water system modelling broadly fall into highly 
+numerical models that excel in representing individual subsystems, or systems 
+dynamics models that create broad representations but that lack a physical 
+basis. Early attempts at a physical representation of the water cycle combined 
+existing numerical models through an integration framework[@Rauch:2017]. While 
+successful, this approach has an incredibly high user burden because each 
+subsystem model is so detailed, and as a consequence is also difficult to 
+customise. To illustrate, SWAT is one of the most widespread models of the 
+rural water cycle[@Arnold:2012], while SWMM is the same but for the urban 
+water cycle[@Gironas:2010]. It has been demonstrated that these two software 
+can interface using the OpenMI integration framework[@Shrestha:2013]. Despite 
+this seemingly powerful combination of two near-ubiquitous models, integrated 
+applications have been limited, and we propose that this is for the same 
+reasons presented in[@Rauch:2017]: user burden and customisation difficulty. 
+
+Because of this need, we provide a parsimonious and self-contained suite for integrated water cycle modelling in the WSIMOD Python package. It 
 brings together a range of software developed over the course of three years 
 on the [CAMELLIA project](https://www.camelliawater.org/). Urban water 
-processes are based on those presented and validated in the CityWat model 
-[@Dobson:2020;@Dobson:2021;@Dobson:2022;@Muhandes:2022], while hydrological 
-and agricultural processes are from the CatchWat model [@Liu:2022;@Liu:2022b]. 
-WSIMOD also provides an interface for message passing between different 
-model components, enabling all parts of the water cycle to interact with all 
-other parts. The result is a simulation model that is easy to set up, highly 
-flexible and ideal for representing water quality and quantity in 'non-
-textbook' water systems (which in our experience is nearly all of them). 
+processes are based on those presented and validated in the CityWat 
+model[@Dobson:2020;@Dobson:2021;@Dobson:2022;@Muhandes:2022], while
+hydrological and agricultural processes are from the CatchWat 
+model[@Liu:2022;@Liu:2022b]. WSIMOD also provides an interface for message
+passing between different model components, enabling all parts of the water 
+cycle to interact with all other parts. The result is a simulation model that
+is easy to set up, highly flexible and ideal for representing water quality and
+quantity in 'non-textbook' water systems (which in our experience is nearly 
+all of them). 
 
 The package provides a variety of tutorials and examples to help modellers 
 create nodes (i.e., representations of subsystems within the water cycle), 
 connect them together with arcs (i.e., representing the fluxes between 
 subsystems), and orchestrate them into a model that creates simulations. 
 
+## Limitations
+We highlight that WSIMOD is not intended to be a substitute for sophisticated 
+physical models, nor for a system dynamics approach. In applications where 
[…]
