+++ setup.py
@@ -1,5 +1,15 @@
 import setuptools
 
+
+with open("requirements.txt") as f:
+    install_requires = f.read().splitlines()
+
+with open("requirements_demos.txt") as f:
+    demos_requires = f.read().splitlines()
+
+with open("requirements_documentation.txt") as f:
+    docs_requires = f.read().splitlines() + demos_requires
+
 setuptools.setup(
    name='WSIMOD',
    version='0.2',
@@ -9,31 +19,9 @@ setuptools.setup(
    license='LICENSE',
    description='WSIMOD is for simulating water quality and quantity',
    long_description=open('README.md').read(),
-   install_requires=['tqdm','pytest','PyYAML','pyarrow','fastparquet'],
+   install_requires=install_requires,
    extras_require={
-        'demos': [ 
-            'pandas', 
-            'geopandas', 
-            'matplotlib', 
-            'shapely', 
-            ],
-      'documentation': 
-      [
-        'mkdocs',
-        'mkdocs-material',
-        'mkdocs-autorefs',
-        'mkdocs-bibtex',
-        'mkdocs-coverage',
-        'mkdocs-jupyter',
-        'mkdocs-material-extensions',
-        'mkdocstrings',
-        'mkdocstrings-python',
-        'pypandoc',
-        'pandas', 
-        'geopandas', 
-        'matplotlib', 
-        'shapely', 
-        'ipykernel',
-      ]
+        'demos': demos_requires,
+      'documentation': docs_requires
     }
       )
