+++ requirements_demos.txt
@@ -1,0 +1,25 @@
+certifi==2022.12.7
+click==8.1.3
+click-plugins==1.1.1
+cligj==0.7.2
+contourpy==1.0.7
+cramjam==2.6.2
+cycler==0.11.0
+fastparquet==2023.1.0
+Fiona==1.9.0
+fonttools==4.38.0
+fsspec==2023.1.0
+geopandas==0.12.2
+kiwisolver==1.4.4
+matplotlib==3.6.3
+munch==2.5.0
+numpy==1.24.1
+pandas==1.5.3
+Pillow==9.4.0
+pyarrow==11.0.0
+pyparsing==3.0.9
+pyproj==3.4.1
+python-dateutil==2.8.2
+pytz==2022.7.1
+shapely==2.0.1
+six==1.16.0
