+++ requirements_documentation.txt
@@ -1,0 +1,79 @@
+asttokens==2.2.1
+backcall==0.2.0
+beautifulsoup4==4.11.1
+bleach==6.0.0
+charset-normalizer==3.0.1
+comm==0.1.2
+debugpy==1.6.6
+decorator==5.1.1
+defusedxml==0.7.1
+entrypoints==0.4
+executing==1.2.0
+fastjsonschema==2.16.2
+ghp-import==2.1.0
+griffe==0.25.4
+idna==3.4
+importlib-metadata==6.0.0
+importlib-resources==5.10.2
+ipykernel==6.20.2
+ipython==8.9.0
+jedi==0.18.2
+Jinja2==3.1.2
+jsonschema==4.17.3
+jupyter_client==8.0.2
+jupyter_core==5.2.0
+jupyterlab-pygments==0.2.2
+jupytext==1.14.4
+kiwisolver==1.4.4
+latexcodec==2.0.1
+lxml==4.9.2
+Markdown==3.3.7
+markdown-it-py==2.1.0
+MarkupSafe==2.1.2
+matplotlib-inline==0.1.6
+mdit-py-plugins==0.3.3
+mdurl==0.1.2
+mergedeep==1.3.4
+mistune==0.8.4
+mkdocs==1.4.2
+mkdocs-autorefs==0.4.1
+mkdocs-bibtex==2.8.10
+mkdocs-coverage==0.2.6
+mkdocs-jupyter==0.22.0
+mkdocs-material==8.5.11
+mkdocs-material-extensions==1.1.1
+mkdocstrings==0.20.0
+mkdocstrings-python==0.8.3
+nbclient==0.7.2
+nbconvert==6.5.4
+nbformat==5.7.3
+nest-asyncio==1.5.6
+pandocfilters==1.5.0
+parso==0.8.3
+pickleshare==0.7.5
+pkgutil_resolve_name==1.3.10
+platformdirs==2.6.2
+prompt-toolkit==3.0.36
+psutil==5.9.4
+pure-eval==0.2.2
+pybtex==0.24.0
+Pygments==2.14.0
+pymdown-extensions==9.9.2
+pypandoc==1.10
+pyrsistent==0.19.3
+pywin32==305
+pyyaml_env_tag==0.1
+pyzmq==25.0.0
+requests==2.28.2
+soupsieve==2.3.2.post1
+stack-data==0.6.2
+tinycss2==1.2.1
+toml==0.10.2
+tornado==6.2
+traitlets==5.9.0
+urllib3==1.26.14
+validators==0.20.0
+watchdog==2.2.1
+wcwidth==0.2.6
+webencodings==0.5.1
+zipp==3.12.0
