+++ docs/installation.md
@@ -14,26 +14,30 @@ Install WSIMOD directly from GitHub
 pip install https://github.com/barneydobson/wsi/archive/refs/heads/main.zip
 ```
 
-Use `[demos]` to include the demos and tutorials.
-```
-pip install -e https://github.com/barneydobson/wsi/archive/refs/heads/main.zip[demos]
-```
-
 If you want to make changes WSIMOD you can download/clone this folder, navigate to it, and run:
 ```
 python setup.py develop
 ```
 
-or (with demos)
+## Run demos
+
+Use `[demos]` to include the demos and tutorials.
+```
+pip install -e https://github.com/barneydobson/wsi/archive/refs/heads/main.zip[demos]
+```
+
+or (with develop)
 
 ```
 python setup.py develop easy_install "wsimod[demos]"
 ```
 
-If you want to compile new documentation you will need to install some additional packages
+## Create documentation
+
+If you want to compile new documentation you will need to install some additional packages:
 
 ```
-python setup.py develop easy_install "wsimod[documentation]"
+pip install -e https://github.com/barneydobson/wsi/archive/refs/heads/main.zip[documentation]
 ```
 
 And then open python, and run:
