### 2023-02-02 e293dde
Create component library page
-Iterates over classes with assumptions in the docstring and writes these to a markdown table
-Update other documentation to reflect this
