+++ requirements_documentation.txt
@@ -67,6 +67,7 @@ pyzmq==25.0.0
 requests==2.28.2
 soupsieve==2.3.2.post1
 stack-data==0.6.2
+tabulate==0.9.0
 tinycss2==1.2.1
 toml==0.10.2
 tornado==6.2
