### 2023-02-02 c039325
Update mkdocs.yml
avoid duplicating component library
