### 2023-02-03 33d740e
Combine push_check_accept into nodes.py
Since currently Rivers now use it because more conceptual work needs to be done to determine river capacity
