### 2023-02-03 606d28b
Update create_class_page.py
-Add nutrient pool
-Fix markdown compatibility issue
