### 2023-02-03 89f90a1
Update demand.py
tidy and update documentation
