### 2023-02-03 64899ca
Update installation.md
(now in contributing)
