### 2023-02-04 f24425b
Update CONTRIBUTING.md
Explain locally compiling documentation
