### 2023-02-06 9440756
Update model.py
add irrigation
