+++ requirements_documentation.txt
@@ -4,6 +4,7 @@ beautifulsoup4==4.11.1
 bleach==6.0.0
 charset-normalizer==3.0.1
 comm==0.1.2
+coverage==7.1.0
 debugpy==1.6.6
 decorator==5.1.1
 defusedxml==0.7.1
