+++ paper/paper.bib
@@ -1,5 +1,6 @@
 @article{Dobson:2020,
   url={https://doi.org/10.1088/1748-9326/abb050},
+  doi={10.1088/1748-9326/abb050},
   title={Protecting rivers by integrating supply-wastewater infrastructure planning and coordinating operational decisions},
   author={Dobson, Barnaby and Mijic, Ana},
   journal={Environmental Research Letters},
@@ -12,6 +13,7 @@
 
 @article{Liu:2022,
   url={https://doi.org/10.1016/j.scitotenv.2021.150642},
+  doi={10.1016/j.scitotenv.2021.150642},
   title={Hierarchical systems integration for coordinated urban-rural water quality management at a catchment scale},
   author={Liu, Leyang and Dobson, Barnaby and Mijic, Ana},
   journal={Science of The Total Environment},
@@ -23,6 +25,7 @@
 
 @article{Rauch:2017,
   url={https://doi.org/10.1016/j.watres.2017.09.039},
+  doi={10.1016/j.watres.2017.09.039},
   title={Modelling transitions in urban water systems},
   author={Rauch, W and Urich, C and Bach, PM and Rogers, BC and De Haan, FJ and Brown, RR and Mair, M and McCarthy, DT and Kleidorfer, M and Sitzenfrei, R and others},
   journal={Water Research},
@@ -34,6 +37,7 @@
 
 @article{Dobson:2021,
   url={https://doi.org/10.3389/frwa.2021.641462},
+  doi={10.3389/frwa.2021.641462},
   title={Integrated modelling to support analysis of COVID-19 impacts on London's water system and in-river water quality},
   author={Dobson, Barnaby and Jovanovic, Tijana and Chen, Yuting and Paschalis, Athanasios and Butler, Adrian and Mijic, Ana},
   journal={Frontiers in Water},
@@ -45,6 +49,7 @@
 
 @article{Dobson:2022,
   url={https://doi.org/10.1029/2021WR030778}, 
+  doi={10.1029/2021WR030778},
   title={A reduced complexity model with graph partitioning for rapid hydraulic assessment of sewer networks},
   author={Dobson, Barnaby and Watson-Hill, Hannah and Muhandes, Samer and Borup, Morten and Mijic, Ana},
   journal={Water Resources Research},
@@ -57,6 +62,7 @@
 
 @article{Muhandes:2022,
   url = {https://doi.org/10.3389/frwa.2021.773974},
+  doi={10.3389/frwa.2021.773974},
   title={The value of aggregated city scale models to rapidly assess SuDS in combined sewer systems},
   author={Muhandes, Samer and Dobson, Barnaby and Mijic, Ana},
   journal={Frontiers in Water},
@@ -66,9 +72,48 @@
 }
 
 @article{Liu:2022b,
-  url = {https://www.essoar.org/doi/abs/10.1002/essoar.10511984.1},
+  url = {https://www.sciencedirect.com/science/article/pii/S0301479722026184},
+  doi={10.1016/j.jenvman.2022.117045},
   title={Optimisation of Urban-Rural Nature-Based Solutions for Integrated Catchment Water Management},
   author={Liu, Leyang and Dobson, Barnaby and Mijic, Ana},
-  year={2022},
-  journal={ESSOAr: Earth and Space Science Open Archive},
+  year={2023},
+  journal={Journal of Environmental Management},
+}
+
+@article{Shrestha:2013,
+  url = {https://www.sciencedirect.com/science/article/pii/S136481521300114X},
+  doi = {10.1016/j.envsoft.2013.05.004},
+  title={OpenMI-based integrated sediment transport modelling of the river Zenne, Belgium},
[…]
