### 2023-02-06 8b91525
Update documentation
-Deploy instructions
-Link contributing and CoC in readme
