+++ docs/demo/scripts/oxford_demo.py
@@ -4,7 +4,7 @@
 # Note - this script can also be opened in interactive Python if you wanted to
 # play around. On the GitHub it is in [docs/demo/scripts](https://github.com/barneydobson/wsi/blob/main/docs/demo/scripts/oxford_demo.py)
 #
-# 1. [Introduction](#we-will-cover-a-demo-WSIMOD-case-study)
+# 1. [Introduction](#we-will-cover-a-demo-wsimod-case-study)
 #
 # 2. [Data](#imports-and-forcing-data)
 #
