### 2023-02-06 5119fe5
Update installation.md
in documentation to match README
