+++ README.md
@@ -73,13 +73,13 @@ pip install -e https://github.com/barneydobson/wsi/archive/refs/heads/main.zip[d
 
 If you want to make changes WSIMOD you can download/clone this folder, navigate to it, and run:
 ```
-pip install .
+python setup.py develop
 ```
 
 or (with demos)
 
 ```
-pip install -e .[demos]
+python setup.py develop easy_install "wsimod[demos]"
 ```
 
 ## Acknowledgements
