### 2023-02-06 913b37f
Update install instructions for demos
