### 2023-02-06 1f0d770
Update land_demo.py
confusion around surface type in documentation
