### 2023-02-06 barneydobson/wsi#7 @barneydobson
Hi @cheginit , thanks so much for the review, it's been incredibly helpful for us. Below I will go through the open tasks and provide links to where they have been addressed and add text for any responses:

- Paper (literature)
   -  No discussion on the current state of the field and the knowledge gap that this package is aiming to fill.
   -  The discussion on the software capabilities are very limited. You need to provide more details on high-level functionalities of the software, such as, Urban processes that it can model, required input data, number of calibrated vs non-calibrated parameters, and its applicability range and limitations.

_Addressed in:
https://github.com/barneydobson/wsi/blob/main/docs/paper/paper.md
https://barneydobson.github.io/wsi/paper/paper/
Note that this is a reasonably small amount of text added, we are currently writing up a full research paper with a more complete explanation of theory. Although RE capabilities, more information provided below_

- Including a (long) TOC in the README is not helpful. It'd be better if it includes contents similar to the [About](https://barneydobson.github.io/wsi/paper/paper/) section of the software's website and some figures demonstrating its capabilities. Essentially, it should give users a quick glance to the content of the repository.
_Updated https://github.com/barneydobson/wsi/blob/main/README.md_

- Using linters could help improve the code quality and catch possible errors. For example, [pre-commit](https://pre-commit.com/) can be a good starting point.
_New issue, to be addressed in future: https://github.com/barneydobson/wsi/issues/9_

- Input data
   -  In my opinion, the way that input data are structured and users have to make all components manually is not user-friendly at all. Since there are many components that required user input, using a config file is much more user-friendly. Many modeling packages follow this input structure, such as SWMM. It's easy to implement such a config file-based input using yaml files in Python.
   -  There is no clear and direct guidance or discussion in the documentation regarding the required input data and pre-processing them in a form that the model accepts. There are some hints in the examples, but there should be a dedicated section for processing raw input data from different formats to the required format.

_Addressed as specified in:
https://github.com/barneydobson/wsi/blob/main/wsimod/orchestration/model.py
https://barneydobson.github.io/wsi/wsimod_models/ (tutorial)
https://barneydobson.github.io/wsi/component-library/ (data requirements)_

-  Considering that this package combines several models, at least a brief discussion on their assumptions and governing equations are required so a user can make informed decisions regarding the necessary components that need to be included for getting acceptable results. A very detailed and technical discussion is not required like a journal paper, but at least a user should be able to have a good understanding of the model assumptions by reading the documentation.
_Addressed as discussed above: https://barneydobson.github.io/wsi/component-library/_
