+++ docs/demo/scripts/land_demo.py
@@ -57,7 +57,7 @@ from matplotlib import pyplot as plt
 data_folder= os.path.join(os.path.abspath(''),
                                "docs","demo","data")
 #Use this path otherwise
-#data_folder = os.path.join(os.path.split(os.path.abspath(''))[0],"data") 
+data_folder = os.path.join(os.path.split(os.path.abspath(''))[0],"data") 
 
 input_fid = os.path.join(data_folder, "processed", "timeseries_data.csv")
 input_data = pd.read_csv(input_fid)
