### 2023-02-07 a1063d4
Update model.py
Enable reservoirs to lie on river network
