+++ wsimod/arcs/arcs.py
@@ -775,7 +775,46 @@ class DecayArcAlt(AltQueueArc, DecayObj):
                 self.queue[i+1] = self.empty_vqip()
 
         self.queue[0] = self.sum_vqip(self.queue[0], self.make_decay(queue_[0]))
-    
+
+class PullArc(Arc):
+    def __init__(self, **kwargs):
+        """Subclass of Arc where pushes return no availability to push. This creates
+        an Arc where only pull requests/checks can be sent, similar to a river
+        abstraction.
+        """
+        super().__init__(**kwargs)
+        self.send_push_request = self.send_push_deny
+        self.send_push_check = self.send_push_check_deny
+    def send_push_deny(self, vqip, tag = 'default', force = False):
+        """Function used to deny any push requests.
+
+        Args:
+            vqip (dict): A dict VQIP of water to push
+            tag (str, optional):  optional message to direct the out_port's query_handler which 
+                function to call. Defaults to 'default'.
+            force (bool, optional): Argument used to cause function to ignore tank 
+                capacity of out_port, possibly resulting in pooling. Should not be used unless 
+                out_port is a tank object. Defaults to False.
+
+        Returns:
+            (dict): A VQIP amount of water that was not successfully pushed
+        """
+        return vqip
+
+    def send_push_check_deny(self, vqip = None, tag = 'default'):
+        """Function used to deny any push checks.
+
+        Args:
+            vqip (dict): A dict VQIP of water to push that can be specified. Defaults to None, 
+                which returns maximum capacity to push.
+            tag (str, optional):  optional message to direct the out_port's query_handler which 
+                function to call. Defaults to 'default'.
+
+        Returns:
+            (dict): An empty VQIP amount of water indicating no water can be pushed
+        """
+        return self.empty_vqip()
+
 class SewerArc(Arc):
     pass
 
