### 2023-02-09 6099e5f
Add self.riverrc as a state variable
To solve the issue of 'Maxiter reached' when pull water from river(storage). Now riverrc won't change within each timestep.
