### 2023-02-19 barneydobson/wsi#7 @cheginit
Thanks for taking the time to address the issues. For the most part, you have addressed my concerns, there's still room for improvement, but in my opinion, the package is useful at its current state.
