### 2023-02-21 barneydobson/wsi#6 @jlarsen-usgs
@barneydobson thank you for addressing my comments. I think this is ready to move ahead
