### 2023-02-21 ba8febe
Update paper.md
Update reference locations
Add examples in limitations
Remove markdown bibliography
