+++ docs/paper/paper.md
@@ -30,9 +30,9 @@ physical and human processes throughout. For example, rivers are a water
 supply, a receiver of wastewater, and an aggregate of many hydrological, 
 biological, and chemical processes. Thus, simulations of the water cycle that
 have highly constrained boundaries may miss key interactions that create 
-unanticipated impacts or unexpected opportunities[@Dobson:2020;@Liu:2022]. 
+unanticipated impacts or unexpected opportunities [@Dobson:2020;@Liu:2022]. 
 Integrated environmental models aim to resolve the issue of boundary 
-conditions, however they have some key limitations[@Rauch:2017], and in 
+conditions, however they have some key limitations [@Rauch:2017], and in 
 particular we find a significant need for a parsimonious, self-contained suite 
 that is accessible and easy to setup. 
 
@@ -42,24 +42,24 @@ Traditional approaches to water system modelling broadly fall into highly
 numerical models that excel in representing individual subsystems, or systems 
 dynamics models that create broad representations but that lack a physical 
 basis. Early attempts at a physical representation of the water cycle combined 
-existing numerical models through an integration framework[@Rauch:2017]. While 
+existing numerical models through an integration framework [@Rauch:2017]. While 
 successful, this approach has an incredibly high user burden because each 
 subsystem model is so detailed, and as a consequence is also difficult to 
 customise. To illustrate, SWAT is one of the most widespread models of the 
-rural water cycle[@Arnold:2012], while SWMM is the same but for the urban 
-water cycle[@Gironas:2010]. It has been demonstrated that these two software 
-can interface using the OpenMI integration framework[@Shrestha:2013]. Despite 
+rural water cycle [@Arnold:2012], while SWMM is the same but for the urban 
+water cycle [@Gironas:2010]. It has been demonstrated that these two software 
+can interface using the OpenMI integration framework [@Shrestha:2013]. Despite 
 this seemingly powerful combination of two near-ubiquitous models, integrated 
 applications have been limited, and we propose that this is for the same 
-reasons presented in[@Rauch:2017]: user burden and customisation difficulty. 
+reasons presented in [@Rauch:2017]: user burden and customisation difficulty. 
 
 Because of this need, we provide a parsimonious and self-contained suite for integrated water cycle modelling in the WSIMOD Python package. It 
 brings together a range of software developed over the course of three years 
 on the [CAMELLIA project](https://www.camelliawater.org/). Urban water 
 processes are based on those presented and validated in the CityWat 
-model[@Dobson:2020;@Dobson:2021;@Dobson:2022;@Muhandes:2022], while
+model [@Dobson:2020;@Dobson:2021;@Dobson:2022;@Muhandes:2022], while
 hydrological and agricultural processes are from the CatchWat 
-model[@Liu:2022;@Liu:2022b]. WSIMOD also provides an interface for message
+model [@Liu:2022;@Liu:2022b]. WSIMOD also provides an interface for message
 passing between different model components, enabling all parts of the water 
 cycle to interact with all other parts. The result is a simulation model that
 is easy to set up, highly flexible and ideal for representing water quality and
@@ -74,9 +74,10 @@ subsystems), and orchestrate them into a model that creates simulations.
 ## Limitations
 We highlight that WSIMOD is not intended to be a substitute for sophisticated 
 physical models, nor for a system dynamics approach. In applications where 
-detailed hydraulic/hydrological process representations are (e.g.,) needed to 
-inform design, or in cases where capturing (e.g.,) hard to quantify social 
-processes, there are likely better tools available. Our case studies highlight 
+detailed hydraulic/hydrological process representations are needed, for example, to 
+inform the design of specific pipes, or in cases where processes are hard to quantify, 
[…]
