+++ docs/paper/paper.md
@@ -11,6 +11,7 @@ authors:
     orcid: 0000-0002-0149-4124
     affiliation: 1
   - name: Leyang Liu
+    orcid: 0000-0001-7556-1134
     affiliation: 1
   - name: Ana Mijic
     orcid: 0000-0001-7096-9405
