### 2023-02-21 #4996 @barneydobson
Thanks @crvernon , those paper changes now corrected.

I've made a release, created an archive with the correct metadata.

With DOI: [10.5281/zenodo.7662569](https://doi.org/10.5281/zenodo.7662569)
