### 2023-02-21 #4996 @crvernon
@barneydobson I just noticed that part of the DOI is missing for the following:

```
Gironás, J., Roesner, L. A., Rossman, L. A., & Davis, J. (2010). A new applications manual
82 for the storm water management model(SWMM). Environmental Modelling & Software,
83 25(6), 813–814. https://doi.org/j.envsoft.2009.11.009
```

The DOI link should be:  https://doi.org/10.1016/j.envsoft.2009.11.009

Also, do mind changing the following sentence in the "Limitations" section **from**:

> In applications where detailed hydraulic/hydrological process representations are needed, for example, to inform the design of specific pipes, or in cases where processes are hard to quantify, for example, in representing social drivers of population growth, there are likely better tools available.

**to**

> In applications where detailed hydraulic/hydrological process representations are needed (e.g., informing the design of specific pipes, cases where processes are hard to quantify such as representing social drivers of population growth, etc.) there are likely better tools available.

That should be the last of it.  Thanks!
