### 2023-02-23 96ca779
Paper updates for JOSS
