+++ docs/paper/paper.bib
@@ -107,7 +107,7 @@
 
 @article{Gironas:2010,
   url = {https://www.sciencedirect.com/science/article/pii/S1364815209002989},
-  doi = {j.envsoft.2009.11.009},
+  doi = {10.1016/j.envsoft.2009.11.009},
   title={A new applications manual for the Storm Water Management Model(SWMM)},
   author={Giron{\'a}s, Jorge and Roesner, Larry A and Rossman, Lewis A and Davis, Jennifer},
   journal={Environmental Modelling \& Software},
