### 2023-02-27 #4996 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.

Also the acknowledgements for "The design of ..." look a little strange — please make sure they are coming through as you intend.
