+++ docs/paper/paper.bib
@@ -117,3 +117,59 @@
   year={2010},
   publisher={Elsevier Science, P. O. Box 800 Kidlington Oxford OX 5 1 DX UK}
 }
+
+@article{Belete:2017,
+  title={An overview of the model integration process: From pre-integration assessment to testing},
+  doi = {10.1016/j.envsoft.2016.10.013},
+  author={Belete, Getachew F and Voinov, Alexey and Laniak, Gerard F},
+  journal={Environmental modelling \& software},
+  volume={87},
+  pages={49--63},
+  year={2017},
+  publisher={Elsevier}
+}
+
+@misc{smif_software,
+      author       = {Will Usher and Tom Russell and Roald Schoenmakers and Craig Robson and Fergus Cooper and Thibault Lestang and Rose Dickinson},
+      title        = {nismod/smif v1.3.2},
+      month        = Aug,
+      year         = 2018,
+      doi          = {10.5281/zenodo.1309336},
+      url          = {https://doi.org/10.5281/zenodo.1309336}
+}
+
+@misc{smif_paper,
+      author       = {Will Usher and Tom Russell},
+      title        = {A Software Framework for the Integration of Infrastructure Simulation Models},
+      journal      = {Journal of Open Research Software},
+      volume       = {7},
+      number       = {16},
+      pages        = {1--5},
+      month        = May,
+      year         = {2019},
+      doi          = {10.5334/jors.265},
+      url          = {https://doi.org/10.5334/jors.265}
+}
+
+@article{Burger:2016,
+  title={Designing and implementing a multi-core capable integrated urban drainage modelling Toolkit: Lessons from CityDrain3},
+  author={Burger, Gregor and Bach, Peter M and Urich, Christian and Leonhardt, G{\"u}nther and Kleidorfer, Manfred and Rauch, Wolfgang},
+  doi = {10.1016/j.advengsoft.2016.08.004},
+  journal={Advances in Engineering Software},
+  volume={100},
+  pages={277--289},
+  year={2016},
+  publisher={Elsevier}
+}
+
+@article{Gregersen:2007,
+  title={OpenMI: Open modelling interface},
+  doi = {10.2166/hydro.2007.023},
+  author={Gregersen, JB and Gijsbers, PJA and Westen, SJP},
+  journal={Journal of hydroinformatics},
+  volume={9},
+  number={3},
+  pages={175--191},
+  year={2007},
+  publisher={IWA Publishing}
+}
