### 2023-02-27 #4996 @barneydobson
@kthyng Thanks for your comments there. I have updated the acknowledgements to include them as references, which I should have done from the start - hopefully you'll agree they are clearer now.

RE capitalizations, they seem OK to me (and are in curly braces in the .bib file) - is there any specific that seem wrong?
