### 2023-03-01 #4996 @kthyng
@barneydobson References comments:

- [ ] Shrestha et al has "belgium" and the river not capitalized
- [ ] Also Gironas needs a space before the acronym in the title
