+++ docs/paper/paper.bib
@@ -83,7 +83,7 @@
 @article{Shrestha:2013,
   url = {https://www.sciencedirect.com/science/article/pii/S136481521300114X},
   doi = {10.1016/j.envsoft.2013.05.004},
-  title={OpenMI-based integrated sediment transport modelling of the river Zenne, Belgium},
+  title={OpenMI-based integrated sediment transport modelling of the {River Zenne, Belgium}},
   author={Shrestha, Narayan Kumar and Leta, Olkeba Tolessa and De Fraine, Bruno and Van Griensven, Ann and Bauwens, Willy},
   journal={Environmental Modelling \& Software},
   volume={47},
@@ -108,7 +108,7 @@
 @article{Gironas:2010,
   url = {https://www.sciencedirect.com/science/article/pii/S1364815209002989},
   doi = {10.1016/j.envsoft.2009.11.009},
-  title={A new applications manual for the Storm Water Management Model(SWMM)},
+  title={A new applications manual for the Storm Water Management Model (SWMM)},
   author={Giron{\'a}s, Jorge and Roesner, Larry A and Rossman, Lewis A and Davis, Jennifer},
   journal={Environmental Modelling \& Software},
   volume={25},
